+++ src/ztbus/cleaning/speed.py
@@ -32,7 +32,7 @@ def clean_speed(df: pl.DataFrame, cfg: SpeedConfig) -> pl.DataFrame:
 
     # Window in samples ≈ window_seconds × sample rate. ZTBus is 1 Hz nominal,
     # so window_seconds ≈ samples. We keep an odd window for centered medians.
-    window = max(3, int(round(cfg.smoothing.window_seconds)))
+    window = max(3, round(cfg.smoothing.window_seconds))
     if window % 2 == 0:
         window += 1
 
@@ -45,19 +45,21 @@ def clean_speed(df: pl.DataFrame, cfg: SpeedConfig) -> pl.DataFrame:
     # Clamp small negatives to zero, leave the rest alone
     df = df.with_columns(
         pl.when(pl.col(RAW_COL) < 0)
-          .then(pl.when(pl.col(RAW_COL) >= cfg.small_negative_threshold_mps)
-                  .then(pl.lit(0.0))
-                  .otherwise(pl.col(RAW_COL)))   # keep the value, but it's flagged
-          .otherwise(pl.col(RAW_COL))
-          .alias("_speed_clamped"),
+        .then(
+            pl.when(pl.col(RAW_COL) >= cfg.small_negative_threshold_mps)
+            .then(pl.lit(0.0))
+            .otherwise(pl.col(RAW_COL))
+        )  # keep the value, but it's flagged
+        .otherwise(pl.col(RAW_COL))
+        .alias("_speed_clamped"),
     )
 
     # Smooth: rolling median (kills spikes) → rolling mean (residual smoothing)
     df = df.with_columns(
         pl.col("_speed_clamped")
-            .rolling_median(window_size=window, min_samples=1, center=True)
-            .rolling_mean(window_size=window, min_samples=1, center=True)
-            .alias(SMOOTHED_COL),
+        .rolling_median(window_size=window, min_samples=1, center=True)
+        .rolling_mean(window_size=window, min_samples=1, center=True)
+        .alias(SMOOTHED_COL),
     ).drop("_speed_clamped")
 
     return df
